requirements.txt
Django==4.2.4
djangorestframework==3.14.0
redis==4.6.0
